Refactor LaserNode class and enhance Laser firing logic
This commit includes several changes to the `LaserNode` class to enhance the laser firing logic and improve the code structure:

1. Restructured the `fire()` method to now return a detailed response instead of just a boolean value. This gives more information about why the laser may not have fired, such as if it's already firing, on cool-down, etc.

2. The `fireOffTimer` and `fireCooldownTimer` are now canceled before being added to the executor, ensuring they're reset before re-use.

3. Removed the redundant `fireCallback()` method as same is being achieved inside the refactored `fire()` method.

4. The method signature of the `setLoopCallback()` has been changed to take in a pointer to the `LaserNode` implementation instead of a copy. Thus reducing overhead.

5. The previous `fireService` and `setLoopService` initialization are replaced with `rclc_service_init_default`, changing from _best_effort_ to _default_ service type.

The above changes make the logic of handling the state of laser and service responses much clearer. They add more meaning to the actions being taken while firing the laser or setting the loop state and establish a better base for future extension and maintenance.